requirements.txt
kafka-python==2.2.0
pandas==2.3.0
numpy==1.26.4
python-dateutil==2.9.0.post0
pytz==2024.1
